json: Add gtk_json_parser_select_member()
This is modeled after JsonReader.selectName() from the Java Moshi
package, see https://square.github.io/moshi/1.x/moshi/com/squareup/moshi/JsonReader.html#selectName-com.squareup.moshi.JsonReader.Options-

It returns the ID of the current member name in the passed i char array
or -1 if it is not found.

This allows writing code like:
  do {
    switch (gtk_json_parser_select_member (parser, { "foo", "bar", NULL })
    {
    case 0:
      /* handle the "foo" member */
      break;
    case 1:
      /* handle the "bar" member */
      break;
    default:
      /* do nothing or complain or whatever */
      break;
    }
  } while (gtk_json_parser_next (parser));

The impementation so far is the simplest one possible, but could be
significantly improved.